backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
python-dotenv==1.0.1

SQLAlchemy==2.0.42
psycopg2-binary
alembic==1.13.2

requests==2.32.3
beautifulsoup4==4.12.3
dateparser==1.2.0

# File handling and validation
Pillow==10.4.0
Werkzeug==3.0.3
marshmallow==3.22.0

# Authentication and security
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0
